+++ lib/src/internal/constants.dart
@@ -67,6 +67,7 @@ class PMConstants {
   static const String mCancelRequestWithCancelToken =
       'cancelRequestWithCancelToken';
   static const String mCancelAllRequest = 'cancelAllRequest';
+  static const String mGetCloudIdentifiers = 'getCloudIdentifiers';
 
   /// Constant value.
   static const int vDefaultThumbnailSize = 150;